requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.18.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
asteroid-filterbanks==0.4.0
attrs==25.4.0
audeer==2.3.1
audiofile==1.6.0
audioread==3.1.0
audmath==1.4.4
av==15.1.0
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
coloredlogs==15.0.1
colorlog==6.10.1
contourpy==1.3.3
ctranslate2==4.6.3
cycler==0.12.1
decorator==5.2.1
docopt==0.6.2
einops==0.8.1
faster-whisper==1.2.1
filelock==3.20.3
flask==3.1.2
flatbuffers==25.12.19
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.1.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
hyperpyyaml==1.2.3
idna==3.11
itsdangerous==2.2.0
jinja2==3.1.6
joblib==1.5.3
julius==0.2.7
kiwisolver==1.4.9
lazy-loader==0.4
librosa==0.11.0
lightning==2.6.0
lightning-utilities==0.15.2
llvmlite==0.46.0
mako==1.3.10
markdown-it-py==4.0.0
markupsafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
networkx==3.6.1
nltk==3.9.2
numba==0.63.1
numpy==2.0.2
ollama==0.6.1
omegaconf==2.3.0
onnxruntime==1.23.2
optuna==4.7.0
packaging==26.0
pandas==2.2.3
pillow==12.1.0
platformdirs==4.5.1
pooch==1.8.2
primepy==1.3
propcache==0.4.1
protobuf==6.33.4
pyannote-audio==3.4.0
pyannote-core==5.0.0
pyannote-database==5.1.3
pyannote-metrics==3.2.1
pyannote-pipeline==3.0.1
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.12.0
pydub==0.25.1
pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytorch-lightning==2.6.0
pytorch-metric-learning==2.9.0
pytz==2025.2
pyyaml==6.0.3
regex==2026.1.15
requests==2.32.5
rich==14.2.0
ruamel-yaml==0.18.17
ruamel-yaml-clib==0.2.15
rvadfast==0.0.5
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
semver==3.0.4
sentencepiece==0.2.1
setuptools==80.10.1
shellingham==1.5.4
six==1.17.0
sortedcontainers==2.4.0
soundfile==0.13.1
soxr==1.0.0
speechbrain==1.0.3
sqlalchemy==2.0.46
sympy==1.14.0
tabulate==0.9.0
tensorboardx==2.6.4
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.8.0
torch-audiomentations==0.12.0
torch-pitch-shift==1.2.5
torchaudio==2.8.0
torchmetrics==1.8.2
tqdm==4.67.1
transformers==4.57.6
typer==0.21.1
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
urllib3==2.6.3
werkzeug==3.1.5
whisperx==3.7.4
yarl==1.22.0
